Moving a few more files.
