+++ keyboards/tronguylabs/m122_3270/m122_3270.c
@@ -13,4 +13,4 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "tronguylabs.h"
+#include "m122_3270.h"
